AftermintLayer/requirements.txt
pynamodb
pyjwt
aws-lambda-powertools